requirements.txt
nltk==3.4
spacy==2.1.3
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.1.0/en_core_web_sm-2.1.0.tar.gz#egg=en_core_web_sm
bert-serving-server==1.9.1
bert-serving-client==1.9.1
fake_useragent
PyExecjs
orderedset